Change some parameter maxima and their handling
This is in preparation for support of other locales. I wanted custom
number parsing to avoid locale differences, and I decided to lower the
maximum input size to remove the need for overflow checking. The seed
handling now actually takes into account the minimum and maximum as a
result, rather than assuming that all values of RAND_TYPE are valid.
This is better anyway. The maximum intervals were increased to be
effectively infinite.